Move isStandaloneTask to types
