+++ plugin-hrm-form/src/components/case/AddReferral.tsx
@@ -23,8 +23,7 @@ import { updateCase } from '../../services/CaseService';
 import { createFormFromDefinition, disperseInputs, splitInHalf } from '../common/forms/formGenerators';
 import type { DefinitionVersion } from '../common/forms/types';
 import { transformValues } from '../../services/ContactService';
-import { StandaloneITask } from '../StandaloneSearch';
-import type { CustomITask } from '../../types/types';
+import type { CustomITask, StandaloneITask } from '../../types/types';
 import { getConfig } from '../../HrmFormPlugin';
 
 type OwnProps = {
