+++ plugin-hrm-form/src/components/case/ViewReferral.tsx
@@ -10,9 +10,8 @@ import * as RoutingActions from '../../states/routing/actions';
 import { CaseLayout } from '../../styles/case';
 import ActionHeader from './ActionHeader';
 import SectionEntry from '../SectionEntry';
-import { StandaloneITask } from '../StandaloneSearch';
 import { formatName } from '../../utils';
-import type { CustomITask } from '../../types/types';
+import type { CustomITask, StandaloneITask } from '../../types/types';
 
 type OwnProps = {
   task: CustomITask | StandaloneITask;
