+++ plugin-hrm-form/src/states/search/reducer.ts
@@ -2,9 +2,8 @@ import { omit } from 'lodash';
 
 import * as t from './types';
 import { INITIALIZE_CONTACT_STATE, RECREATE_CONTACT_STATE, REMOVE_CONTACT_STATE, GeneralActionType } from '../types';
-import { SearchContact, SearchCaseResult } from '../../types/types';
+import { SearchContact, SearchCaseResult, standaloneTaskSid } from '../../types/types';
 import { ContactDetailsSections, ContactDetailsSectionsType } from '../../components/common/ContactDetails';
-import { standaloneTaskSid } from '../../components/StandaloneSearch';
 
 type PreviousContacts = {
   contacts?: t.DetailedSearchContactsResult;
